Revert

+++ core/queue/base.py
@@ -1,6 +1,7 @@
 import asyncio
 import time
 import traceback
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 from uuid import uuid4
 
@@ -14,12 +15,16 @@ from core.database.models import JobQueuesTable
 from core.exports import exports
 from core.logger import Logger
 
+
 if TYPE_CHECKING:
     from core.builtins.bot import Bot
 
 
-class QueueFinished(Exception):
-    pass
+@dataclass
+class QueueFinished:
+    task_id: str
+    action: str
+    status: str
 
 
 class QueueTaskManager:
@@ -47,7 +52,6 @@ class QueueTaskManager:
 
 class JobQueueBase:
     name = "Internal|" + str(uuid4())
-    _queue_tasks = {}
     queue_actions = {}
     report_targets = Config("report_targets", [])
     is_running = False
@@ -68,7 +72,7 @@ class JobQueueBase:
     @staticmethod
     async def return_val(tsk: JobQueuesTable, value: dict, status: str = "done"):
         await tsk.set_val(value, status)
-        raise QueueFinished
+        return QueueFinished(str(tsk.task_id), tsk.action, tsk.status)
 
     @classmethod
     async def _process_task(cls, tsk: JobQueuesTable):
@@ -77,42 +81,43 @@ class JobQueueBase:
             timestamp = tsk.timestamp
             if time.time() - timestamp.timestamp() > 7200:
                 Logger.warning(f"Task {tsk.task_id} timeout, skip.")
-                await cls.return_val(tsk, {}, status="timeout")
+                tsk_val = await cls.return_val(tsk, {}, status="timeout")
             elif tsk.action in cls.queue_actions:
                 returns: dict = await cls.queue_actions[tsk.action](tsk, tsk.args)
-                await cls.return_val(tsk, returns if returns else {}, status="done")
+                tsk_val = await cls.return_val(tsk, returns if returns else {})
             else:
                 Logger.warning(f"Unknown action {tsk.action}, skip.")
-                await cls.return_val(tsk, {}, status="failed")
-        except QueueFinished:
-            Logger.trace(f"Task {tsk.action}({tsk.task_id}) finished.")
-            await asyncio.sleep(5)
-            await tsk.delete()
-            return
+                tsk_val = await cls.return_val(tsk, {}, status="failed")
+
+            if tsk_val:
+                Logger.trace(f"Task {tsk.action}({tsk.task_id}) {tsk.status}.")
+
+                # Auto-delete done tasks after 5 seconds to reduce database size.
+                # Assume 5 seconds is enough for processing the possible things after task is done.
+                if tsk.status == "done":
+                    await asyncio.sleep(5)
+                    await tsk.delete()
+                return
+            # The code below should not be reached if the task is processed correctly.
+            Logger.error(f"Task {tsk.action}({tsk.task_id}) seems not finished properly, bug in code?")
+            await tsk.set_status("failed")
         except Exception:
             f = traceback.format_exc()
             Logger.error(f)
-            try:
-                await cls.return_val(tsk, {"traceback": f}, status="failed")
-            except QueueFinished:
-                pass
+            await cls.return_val(tsk, {"traceback": f}, status="failed")
             try:
                 for target in cls.report_targets:
                     if ft := await bot.fetch_target(target):
                         await cls.client_direct_message(ft, MessageChain.assign(
                             [I18NContext("error.message.report", command=tsk.action),
-                             Plain(f.strip(), disable_joke=True)]), enable_parse_message=False,
+                                Plain(f.strip(), disable_joke=True)]), enable_parse_message=False,
                             disable_secret_check=True)
             except Exception:
                 Logger.exception()
             return
 
-        # The code below should not be reached if the task is processed correctly.
-        Logger.error(f"Task {tsk.action}({tsk.task_id}) seems not finished properly, bug in code?")
-        await tsk.set_status("failed")
-
     @classmethod
-    async def _check_queue(cls, target_client: str = None):
+    async def _check_queue(cls, target_client: str | None = None):
         # Logger.debug(f"Checking job queue for {cls.name}, target client: {target_client if target_client else "all"}")
         for task_id in QueueTaskManager.tasks.copy():
             tsk = await JobQueuesTable.get_or_none(task_id=task_id)
@@ -130,7 +135,7 @@ class JobQueueBase:
             asyncio.create_task(cls._process_task(tsk))
 
     @classmethod
-    async def check_job_queue(cls, target_client: str = None):
+    async def check_job_queue(cls, target_client: str | None = None):
         if cls.is_running:
             raise QueueAlreadyRunning
         cls.is_running = True
@@ -157,4 +162,4 @@ class JobQueueBase:
                           {"session_info": converter.unstructure(session_info),
                            "message": converter.unstructure(message, MessageChain | MessageNodes),
                            "enable_parse_message": enable_parse_message,
-                           "disable_secret_check": disable_secret_check}, )
+                           "disable_secret_check": disable_secret_check})
